requirements.txt
anyio==4.3.0
certifi==2024.2.2
charset-normalizer==3.3.2
geojson==2.5.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.6
paho-mqtt==2.1.0
pyowm==3.3.0
PySocks==1.7.1
python-dotenv==1.0.1
python-telegram-bot==21.0.1
requests==2.31.0
schedule==1.2.1
sniffio==1.3.1
urllib3==2.2.1
